move contextvars flag to compat
