+++ ddtrace/internal/context_manager.py
@@ -10,13 +10,6 @@ log = get_logger(__name__)
 
 _DD_CONTEXTVAR = contextvars.ContextVar("datadog_contextvar", default=None)
 
-try:
-    import contextvars
-except ImportError:
-    CONTEXTVARS_IS_AVAILABLE = False
-else:
-    CONTEXTVARS_IS_AVAILABLE = True
-
 
 class BaseContextManager(six.with_metaclass(abc.ABCMeta)):
     def __init__(self, reset=True):
